Cloning the Android SQLite code into its JDBC counterpart. Includes an implementation for the JDBC update tool updater, which was empty.

+++ srm/codegen/java/jdbc-runtime/src/main/java/com/th3l4b/srm/codegen/java/jdbcruntime/DefaultJDBCIdentifierParser.java
@@ -10,7 +10,7 @@ public class DefaultJDBCIdentifierParser implements IJDBCIdentifierParser {
 
 	@Override
 	public IIdentifier parse(Integer column, ResultSet result) throws Exception {
-		return new DefaultIdentifier(result.getString(column));
+		return new DefaultIdentifier(result.getString(column.intValue()));
 	}
 
 	@Override
